Add edge calculations

+++ lib/core/src/synapse.rs
@@ -6,7 +6,7 @@ pub struct Synapse {
     pub is_dead: bool,
     pub error_der: f32,
     pub accumulated_error_der: f32,
-    pub total_accumulated_error_der: f32,
+    pub total_accumulated_error_der: i32,
     pub regularization: Regularization
 }
 
@@ -17,7 +17,7 @@ impl Synapse {
             is_dead: false,
             error_der: 0f32,
             accumulated_error_der: 0f32,
-            total_accumulated_error_der: 0f32,
+            total_accumulated_error_der: 0,
             regularization
         }
     }
